Recognize already highlighted code snippet in all case.
In Writer, when snippet is plain text (no frame, no shape nor table cell) and
no text is selected and the snippet has already been highlighted with CH2,
CH2 will recongize it and select it.
Until now, this was available only for frame, shape, text cell or when calling update.